add a coplanar polygon check and split quads if necessary
The number used to determine how far a vertex is away from the plane
before it is considered not on the plane probably needs adjustment.

Only quads are split into two triangles. The quad is not checked if it
is self intersecting (bow tie).

Concave and self intersecting checks are needed for more complex
polygons.